Finish refactoring the Positron data grid
