+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridCornerTopLeft.css
