+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridRowCell.tsx
@@ -12,10 +12,10 @@ import { MouseEvent } from 'react'; // eslint-disable-line no-duplicate-imports
 // Other dependencies.
 import { isMacintosh } from 'vs/base/common/platform';
 import { positronClassNames } from 'vs/base/common/positronUtilities';
-import { IDataColumn } from 'vs/base/browser/ui/dataGrid/interfaces/dataColumn';
-import { usePositronDataGridContext } from 'vs/base/browser/ui/dataGrid/dataGridContext';
+import { IDataColumn } from 'vs/base/browser/ui/positronDataGrid/interfaces/dataColumn';
+import { usePositronDataGridContext } from 'vs/base/browser/ui/positronDataGrid/positronDataGridContext';
 import { VerticalSplitter } from 'vs/base/browser/ui/positronComponents/verticalSplitter';
-import { CellSelectionState } from 'vs/base/browser/ui/dataGrid/interfaces/dataGridInstance';
+import { CellSelectionState } from 'vs/base/browser/ui/positronDataGrid/interfaces/dataGridInstance';
 import { HorizontalSplitter } from 'vs/base/browser/ui/positronComponents/horizontalSplitter';
 
 /**
