+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridRowHeader.css
