+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridRowHeader.tsx
@@ -11,10 +11,10 @@ import { MouseEvent } from 'react'; // eslint-disable-line no-duplicate-imports
 
 // Other dependencies.
 import { positronClassNames } from 'vs/base/common/positronUtilities';
-import { usePositronDataGridContext } from 'vs/base/browser/ui/dataGrid/dataGridContext';
-import { selectionType } from 'vs/base/browser/ui/dataGrid/utilities/mouseUtilities';
+import { usePositronDataGridContext } from 'vs/base/browser/ui/positronDataGrid/positronDataGridContext';
+import { selectionType } from 'vs/base/browser/ui/positronDataGrid/utilities/mouseUtilities';
 import { VerticalSplitter } from 'vs/base/browser/ui/positronComponents/verticalSplitter';
-import { RowSelectionState } from 'vs/base/browser/ui/dataGrid/interfaces/dataGridInstance';
+import { RowSelectionState } from 'vs/base/browser/ui/positronDataGrid/interfaces/dataGridInstance';
 import { HorizontalSplitter } from 'vs/base/browser/ui/positronComponents/horizontalSplitter';
 
 /**
