+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridRowHeaders.css
