+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridScrollbar.css
