+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridScrollbar.tsx
