+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridScrollbarCorner.css
