+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridScrollbarCorner.tsx
@@ -7,7 +7,7 @@ import 'vs/css!./dataGridScrollbarCorner';
 
 // React.
 import * as React from 'react';
-import { usePositronDataGridContext } from 'vs/base/browser/ui/dataGrid/dataGridContext';
+import { usePositronDataGridContext } from 'vs/base/browser/ui/positronDataGrid/positronDataGridContext';
 
 /**
  * DataGridScrollbarCornerProps interface.
