+++ src/vs/base/browser/ui/positronDataGrid/components/dataGridWaffle.css
