+++ src/vs/base/browser/ui/positronDataGrid/interfaces/dataCell.ts
