+++ src/vs/base/browser/ui/positronDataGrid/positronDataGrid.tsx
@@ -10,8 +10,8 @@ import * as React from 'react';
 
 // Other dependencies.
 import { positronClassNames } from 'vs/base/common/positronUtilities';
-import { DataGridWaffle } from 'vs/base/browser/ui/dataGrid/components/dataGridWaffle';
-import { PositronDataGridConfiguration, PositronDataGridContextProvider } from 'vs/base/browser/ui/dataGrid/dataGridContext';
+import { DataGridWaffle } from 'vs/base/browser/ui/positronDataGrid/components/dataGridWaffle';
+import { PositronDataGridConfiguration, PositronDataGridContextProvider } from 'vs/base/browser/ui/positronDataGrid/positronDataGridContext';
 
 /**
  * DataGridProps interface.
@@ -24,11 +24,11 @@ interface PositronDataGridProps extends PositronDataGridConfiguration {
 }
 
 /**
- * DataGrid component.
- * @param props A DataGridProps that contains the component properties.
+ * PositronDataGrid component.
+ * @param props A PositronDataGridProps that contains the component properties.
  * @returns The rendered component.
  */
-export const DataGrid = (props: PositronDataGridProps) => {
+export const PositronDataGrid = (props: PositronDataGridProps) => {
 	// Render.
 	return (
 		<PositronDataGridContextProvider {...props}>
